[rules] fixed save/reload differences
The problem was that types are converted to more basic forms when saved
to JSON, and JSON requires that all keys are strings. So rule save data
needed to be casted to the proper types after being loaded.

added checks to rule tests to make sure that saved data is loaded the
same.

changed rule_abc tests to save to tmp_path
fixed wrong Error check in rule_ids_test