The relative_entropy column no longer generates infinities.
An Infinity in the relative_entropy column would cause the DNN to have infinite weights during training against that feature set (and crashing), so this possibility was removed.
ID_Entropy now returns an e_relative of 100 if there is a division by zero.
The flood attack in malicious_generators no longer produces decreasing timestamps, which contributed to the problem.